refactor: split llm.ts into focused prompt modules
Split the 794-line llm.ts into:
- llm.ts (152 lines) - Core LLM provider abstraction
- prompts/persona.ts (198 lines) - Agent persona generation
- prompts/deliberation.ts (182 lines) - Session summaries and strategy
- prompts/evaluation.ts (203 lines) - Judge verdicts and next topics
- prompts/interaction.ts (75 lines) - Agent reflections

This improves maintainability by separating concerns and keeping
each file at a manageable size.
